WOM-454 Improve subscription robustness and refactor unsubscribe

See merge request ska-telescope/ska-mid-dish-steering-control!22

+++ src/ska_mid_dish_steering_control/sculib.py
@@ -1549,7 +1549,7 @@ class SteeringControlUnit:
     # --------------------
     # OPC-UA subscriptions
     # --------------------
-    # pylint: disable=dangerous-default-value
+    # pylint: disable=dangerous-default-value,too-many-branches
     def subscribe(
         self,
         attributes: str | list[str],
@@ -1580,6 +1580,8 @@ class SteeringControlUnit:
             ]
         nodes = []
         missing_nodes = []
+        bad_nodes = []
+        uid = time.monotonic_ns()
         for attribute in attributes:
             if attribute != "":
                 if attribute in self.nodes:
@@ -1596,41 +1598,44 @@ class SteeringControlUnit:
             self._client.create_subscription(period, subscription_handler),
             self.event_loop,
         ).result()
-        subscribe_nodes = list(set(nodes))  # Remove any potential node duplicates
-        bad_nodes = []
-        try:
-            handles = asyncio.run_coroutine_threadsafe(
-                subscription.subscribe_data_change(subscribe_nodes), self.event_loop
-            ).result()
-        except ua.UaStatusCodeError as e:
-            # Exceptions are only generated when subscribe_data_change is called with a
-            # single node input.
-            msg = (
-                f"Failed to subscribe to node '{subscribe_nodes[0].nodeid.to_string()}'"
-            )
-            asyncio.run_coroutine_threadsafe(handle_exception(e, msg), self.event_loop)
-            bad_nodes.append(subscribe_nodes[0])
-            subscribe_nodes.pop(0)
+        if nodes:
+            subscribe_nodes = list(set(nodes))  # Remove any potential node duplicates
+            try:
+                handles = asyncio.run_coroutine_threadsafe(
+                    subscription.subscribe_data_change(subscribe_nodes), self.event_loop
+                ).result()
+            except ua.UaStatusCodeError as e:
+                # Exceptions are only generated when subscribe_data_change is called
+                # with a single node input.
+                msg = (
+                    "Failed to subscribe to node "
+                    f"'{subscribe_nodes[0].nodeid.to_string()}'"
+                )
+                asyncio.run_coroutine_threadsafe(
+                    handle_exception(e, msg), self.event_loop
+                )
+                bad_nodes.append(subscribe_nodes[0])
+                subscribe_nodes.pop(0)
 
-        # subscribe_data_change returns an int for a single node, and a list for
-        # mulitple nodes
-        if isinstance(handles, int):
-            handles = list(handles)
-        # The list contains ints for sucessful subscriptions, and status codes when
-        # the subscription has failed.
-        else:
-            for i, node in enumerate(subscribe_nodes):
-                if isinstance(handles[i], ua.uatypes.StatusCode):
-                    bad_nodes.append(node)
-                    subscribe_nodes.pop(i)
-                    handles.pop(i)
+            # subscribe_data_change returns an int for a single node, and a list for
+            # mulitple nodes
+            if isinstance(handles, int):
+                handles = list(handles)
+            # The list contains ints for sucessful subscriptions, and status codes when
+            # the subscription has failed.
+            else:
+                for i, node in enumerate(subscribe_nodes):
+                    if isinstance(handles[i], ua.uatypes.StatusCode):
+                        bad_nodes.append(node)
+                        subscribe_nodes.pop(i)
+                        handles.pop(i)
 
-        uid = time.monotonic_ns()
-        self._subscriptions[uid] = {
-            "handles": handles,
-            "nodes": subscribe_nodes,
-            "subscription": subscription,
-        }
+            self._subscriptions[uid] = {
+                "handles": handles,
+                "nodes": subscribe_nodes,
+                "subscription": subscription,
+            }
+
         return uid, missing_nodes, bad_nodes
 
     def unsubscribe(self, uid: int) -> None:
@@ -1639,21 +1644,8 @@ class SteeringControlUnit:
 
         :param uid: The ID of the user to unsubscribe.
         """
-        subscription = self._subscriptions.pop(uid)
-        try:
-            _ = asyncio.run_coroutine_threadsafe(
-                subscription["subscription"].delete(), self.event_loop
-            ).result()
-        except ConnectionError as e:
-            asyncio.run_coroutine_threadsafe(
-                handle_exception(e, f"Tried to unsubscribe from {uid}."),
-                self.event_loop,
-            )
-
-    def unsubscribe_all(self) -> None:
-        """Unsubscribe all subscriptions."""
-        while len(self._subscriptions) > 0:
-            uid, subscription = self._subscriptions.popitem()
+        subscription = self._subscriptions.pop(uid, None)
+        if subscription is not None:
             try:
                 _ = asyncio.run_coroutine_threadsafe(
                     subscription["subscription"].delete(), self.event_loop
@@ -1663,6 +1655,11 @@ class SteeringControlUnit:
                     handle_exception(e, f"Tried to unsubscribe from {uid}."),
                     self.event_loop,
                 )
+
+    def unsubscribe_all(self) -> None:
+        """Unsubscribe all subscriptions."""
+        for uid in self._subscriptions:
+            self.unsubscribe(uid)
 
     def get_subscription_values(self) -> list[dict]:
         """
